Split snow.rs (713 lines) into snow/ subdirectory module
Reorganize the snow module into a subdirectory with focused sub-files:
- types.rs (157 lines): constants, SnowGrid, SnowPlowingState, SnowStats
- systems.rs (258 lines): pure helper functions and ECS systems
- tests.rs (293 lines): all unit tests
- mod.rs (38 lines): module declarations, re-exports, and SnowPlugin

All public exports preserved via `pub use` in mod.rs so external
callers are unaffected.

Closes #1389
